requirements.txt
certifi==2018.8.13
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
mkl-fft==1.0.4
mkl-random==1.0.1
numpy==1.15.0
oauthlib==2.1.0
openapi-codec==1.3.2
openpyxl==2.5.5
pandas==0.23.4
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
